requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --extra=dev setup.py
#
attrs==21.4.0
    # via
    #   cattrs
    #   jsii
aws-cdk-lib==2.24.1
    # via elkk-stack (setup.py)
awscli==1.24.0
    # via elkk-stack (setup.py)
black==22.3.0
    # via elkk-stack (setup.py)
boto3==1.23.0
    # via elkk-stack (setup.py)
botocore==1.26.0
    # via
    #   awscli
    #   boto3
    #   s3transfer
cattrs==22.1.0
    # via jsii
click==8.1.3
    # via black
colorama==0.4.4
    # via awscli
constructs==10.1.7
    # via
    #   aws-cdk-lib
    #   elkk-stack (setup.py)
docutils==0.15.2
    # via awscli
exceptiongroup==1.0.0rc5
    # via cattrs
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
jsii==1.58.0
    # via
    #   aws-cdk-lib
    #   constructs
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via black
publication==0.0.3
    # via
    #   aws-cdk-lib
    #   constructs
pyasn1==0.4.8
    # via rsa
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
pyyaml==5.4.1
    # via awscli
rsa==4.7.2
    # via awscli
s3transfer==0.5.2
    # via
    #   awscli
    #   boto3
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via black
typing-extensions==4.2.0
    # via
    #   black
    #   jsii
urllib3==1.26.9
    # via botocore

# The following packages are considered to be unsafe in a requirements file:
# pip
